+++ scanpipe/pipes/scancode.py
@@ -39,10 +39,10 @@ from commoncode.resource import VirtualCodebase
 from extractcode import api as extractcode_api
 from packagedcode import get_package_handler
 from packagedcode import models as packagedcode_models
-from scancode import ScancodeError
 from scancode import Scanner
 from scancode import api as scancode_api
 from scancode import cli as scancode_cli
+from scancode.cli import run_scan as scancode_run_scan
 
 from scanpipe import pipes
 from scanpipe.models import CodebaseResource
@@ -428,31 +428,44 @@ def assemble_packages(project):
                     logger.info(f"Unknown Package assembly item type: {item!r}")
 
 
-def run_scancode(location, output_file, options, raise_on_error=False):
+def get_pretty_params(args):
+    """Format provided ``args`` for the ``pretty_params`` run_scan argument."""
+    return {f"--{key.replace('_', '-')}": value for key, value in args.items()}
+
+
+def run_scan(location, output_file, run_scan_args):
     """
     Scan the `location` content and write the results into an `output_file`.
-    The `scancode` executable will run using the provided `options`.
-    If `raise_on_error` is enabled, a ScancodeError will be raised if the
-    exitcode is greater than 0.
+    If `raise_on_error` is enabled, a ScancodeError will be raised if an error occurs
+    during the scan.
     """
-    options_from_settings = settings.SCANCODE_TOOLKIT_CLI_OPTIONS
-    max_workers = get_max_workers(keep_available=1)
+    run_args = settings.SCANCODE_TOOLKIT_RUN_SCAN_ARGS.copy()
+    # The run_scan_args should override any values provided in the settings
+    run_args.update(run_scan_args)
 
-    scancode_args = [
-        pipes.get_bin_executable("scancode"),
-        shlex.quote(location),
-        *options_from_settings,
-        *options,
-        f"--processes {max_workers}",
-        "--verbose",
-        f"--json-pp {shlex.quote(output_file)}",
-    ]
+    if "timeout" in run_args:
+        run_args["timeout"] = int(run_args.get("timeout"))
 
-    exitcode, output = pipes.run_command(scancode_args, log_output=True)
-    if exitcode > 0 and raise_on_error:
-        raise ScancodeError(output)
+    success, results = scancode_run_scan(
+        input=shlex.quote(location),
+        processes=get_max_workers(keep_available=1),
+        quiet=True,
+        verbose=False,
+        return_results=True,
+        echo_func=None,
+        pretty_params=get_pretty_params(run_args),
+        **run_args,
+    )
 
-    return exitcode, output
+    if success:
+        Path(output_file).write_text(json.dumps(results, indent=2))
+        return
+
+    errors = {}
+    for file in results.get("files", []):
+        if scan_errors := file.get("scan_errors"):
+            errors[file.get("path")] = scan_errors
+    return errors
 
 
 def get_virtual_codebase(project, input_location):